refactor existed sortedintlist - split large file for list
